requirements.txt
pytest-playwright
playwright-recaptcha